moved configuration of command line options and config file options to Options class

+++ src/common/Connection.h
@@ -9,7 +9,7 @@
 #include <mutex>
 #include <sstream>
 
-#define DEFAULT_PORT 13477
+#define DEFAULT_PORT ((unsigned short)13477)
 
 namespace asio = boost::asio;
